Fixes ticket #2 (exception & reboot)
see: https://github.com/cmmakerclub/MQTT-Connector/issues/2
I was using this code on an ESP8266 with Mosquitto 1.4.9 (Windows) and the chip was rebooting while it should recover a connection loss.
This fix adresses the reboot issue.

A code review would be appreciated before merging pull request as I didn't dig a lot into the code...