restored uniformity of topics format
Topics were defined this way: /CMMC/<ID>/<topic>

Then in recent versions it turned to:

sometimes /CMMC/<ID>/<topic>,
sometimes /CMMC<ID>/<topic>

So, to be consistent, I put '/' between CMMC and the ID, for every topic
defined by the library, including the lwt topic.

Also, it helps retrocompatibility with existing applications that were
parsing the ID of the board from the topic.